Refactored getWeighted for Weighted nodes to return double instead of Double. Added Averages utility class to contain various algorithms for taking averaged. Moved the averaging methods in ComplexNumber to Averages class.
git-svn-id: svn://svn.syncleus.com/dANN/trunk@425 6ae8b97b-f314-0410-8212-aecf10b92ded

+++ src/com/syncleus/dann/graph/pathfinding/astar/EncapsulatingAstarGraph.java
@@ -87,7 +87,7 @@ public final class EncapsulatingAstarGraph<G extends Graph<? extends Node<? exte
 
 					AstarNode<Node> newDestrination = newNodeMapping.get(destination);
 					if((useGraphWeights)&&(path instanceof WeightedEdge))
-						newSource.connectTo(newDestrination, ((WeightedEdge)path).getWeight().doubleValue());
+						newSource.connectTo(newDestrination, ((WeightedEdge)path).getWeight());
 					else
 						newSource.connectTo(newDestrination, 1.0);
 				}
